Move DAP traits to `traits` module.
These are the most important traits, it's nice to have them together so
it's easy to have a bird's eye view of the API and easily spot inconsistencies.
