+++ probe-rs/src/architecture/arm/mod.rs
@@ -5,13 +5,13 @@ pub(crate) mod core;
 pub mod dp;
 pub mod memory;
 pub mod swo;
+mod traits;
 
 pub use communication_interface::{
-    ApInformation, ArmChipInfo, ArmCommunicationInterface, DapError, MemoryApInformation,
-    RawDapAccess,
+    ApInformation, ArmChipInfo, ArmCommunicationInterface, DapError, MemoryApInformation, Register,
 };
-pub use communication_interface::{PortType, Register};
 pub use swo::{SwoAccess, SwoConfig, SwoMode};
+pub use traits::*;
 
 pub use self::core::m0;
 pub use self::core::m33;
